Disable some Lombok annotations
I want to move away from this ultimately, so we'll do the easy ones for
now.

+++ plugin-api/src/main/java/net/clgd/ccemux/api/rendering/TerminalFont.java
@@ -153,7 +153,7 @@ public abstract class TerminalFont {
 	 */
 	@Nonnull
 	public Rectangle getCharCoords(char c) {
-		int charcode = (int) c;
+		int charcode = c;
 		return new Rectangle(margin + charcode % COLUMNS * (getCharWidth() + margin * 2), margin + charcode / ROWS * (getCharHeight() + margin * 2), getCharWidth(), getCharHeight());
 	}
